+++ xradar/io/backends/gamic.py
@@ -420,11 +420,11 @@ class GamicBackendEntrypoint(BackendEntrypoint):
         Defaults to False, no reindexing. Given dict should contain the kwargs to
         reindex_angle. Only invoked if `decode_coord=True`.
     fix_second_angle : bool
-        For PPI only. If True, fixes erroneous second angle data. Defaults to False.
+        For PPI only. If True, fixes erroneous second angle data. Defaults to ``False``.
     site_coords : bool
         Attach radar site-coordinates to Dataset, defaults to ``True``.
-    kwargs :  kwargs
-        Additional kwargs are fed to `xr.open_dataset`.
+    kwargs : dict
+        Additional kwargs are fed to :py:func:`xarray.open_dataset`.
     """
 
     description = "Open GAMIC HDF5 (.h5, .hdf5, .mvol) using h5netcdf in Xarray"
@@ -520,7 +520,7 @@ class GamicBackendEntrypoint(BackendEntrypoint):
 
 
 def open_gamic_datatree(filename_or_obj, **kwargs):
-    """Open GAMIC HDF5 dataset as xradar Datatree.
+    """Open GAMIC HDF5 dataset as :py:class:`datatree.DataTree`.
 
     Parameters
     ----------
@@ -541,15 +541,15 @@ def open_gamic_datatree(filename_or_obj, **kwargs):
         Defaults to False, no reindexing. Given dict should contain the kwargs to
         reindex_angle. Only invoked if `decode_coord=True`.
     fix_second_angle : bool
-        If True, fixes erroneous second angle data. Defaults to False.
+        If True, fixes erroneous second angle data. Defaults to ``False``.
     site_coords : bool
         Attach radar site-coordinates to Dataset, defaults to ``True``.
-    kwargs :  kwargs
-        Additional kwargs are fed to `xr.open_dataset`.
+    kwargs : dict
+        Additional kwargs are fed to :py:func:`xarray.open_dataset`.
 
     Returns
     -------
-    dtree: DataTree
+    dtree: datatree.DataTree
         DataTree
     """
     # handle kwargs, extract first_dim